issue #99 Warning before switching branches
Why: Running dep update could cause branches to get changed with no
warning. This could cause problems and confusion.

This change addresses the need by:

* The update command does a verify check to ensure all dependencies are
on the correct branch. If they are not the user is given a choice of how
to handle the difference. Would they like to update their config, switch
to the desired branch, or would they like to abort and fix something
themselves.