issue #126 Fixes help <verb>
Why: GitDepend incorrectly displayed the main help page when it should
have been showing help for a specific verb.

This change addresses the need by:

* Works around an issue with CommandLineParser by checking to see if the
verb is help. If it is, it does a quick check to see what verb was
actually being asked for, and passes that to the help system.